move notice code from metamask-controller + config-manager, in to notice-controller
