+++ test/unit/config-manager-test.js
@@ -13,100 +13,6 @@ describe('config-manager', function() {
     configManager = configManagerGen()
   })
 
-  describe('notices', function() {
-    describe('#getNoticesList', function() {
-      it('should return an empty array when new', function() {
-        var testList = [{
-          id:0,
-          read:false,
-          title:"Futuristic Notice"
-        }]
-        var result = configManager.getNoticesList()
-        assert.equal(result.length, 0)
-      })
-    })
-
-    describe('#setNoticesList', function() {
-      it('should set data appropriately', function () {
-        var testList = [{
-          id:0,
-          read:false,
-          title:"Futuristic Notice"
-        }]
-        configManager.setNoticesList(testList)
-        var testListId = configManager.getNoticesList()[0].id
-        assert.equal(testListId, 0)
-      })
-    })
-
-    describe('#updateNoticeslist', function() {
-      it('should integrate the latest changes from the source', function() {
-        var testList = [{
-          id:55,
-          read:false,
-          title:"Futuristic Notice"
-        }]
-        configManager.setNoticesList(testList)
-        configManager.updateNoticesList().then(() => {
-          var newList = configManager.getNoticesList()
-          assert.ok(newList[0].id === 55)
-          assert.ok(newList[1])
-        })
-      })
-      it('should not overwrite any existing fields', function () {
-        var testList = [{
-          id:0,
-          read:false,
-          title:"Futuristic Notice"
-        }]
-        configManager.setNoticesList(testList)
-        configManager.updateNoticesList().then(() => {
-          var newList = configManager.getNoticesList()
-          assert.equal(newList[0].id, 0)
-          assert.equal(newList[0].title, "Futuristic Notice")
-          assert.equal(newList.length, 1)
-        })
-      })
-    })
-
-    describe('#markNoticeRead', function () {
-      it('should mark a notice as read', function () {
-        var testList = [{
-          id:0,
-          read:false,
-          title:"Futuristic Notice"
-        }]
-        configManager.setNoticesList(testList)
-        configManager.markNoticeRead(testList[0])
-        var newList = configManager.getNoticesList()
-        assert.ok(newList[0].read)
-      })
-    })
-
-    describe('#getLatestUnreadNotice', function () {
-      it('should retrieve the latest unread notice', function () {
-        var testList = [
-          {id:0,read:true,title:"Past Notice"},
-          {id:1,read:false,title:"Current Notice"},
-          {id:2,read:false,title:"Future Notice"},
-        ]
-        configManager.setNoticesList(testList)
-        var latestUnread = configManager.getLatestUnreadNotice()
-        assert.equal(latestUnread.id, 2)
-      })
-      it('should return undefined if no unread notices exist.', function () {
-        var testList = [
-          {id:0,read:true,title:"Past Notice"},
-          {id:1,read:true,title:"Current Notice"},
-          {id:2,read:true,title:"Future Notice"},
-        ]
-        configManager.setNoticesList(testList)
-        var latestUnread = configManager.getLatestUnreadNotice()
-        assert.ok(!latestUnread)
-      })
-    })
-  })
-
   describe('currency conversions', function() {
 
     describe('#getCurrentFiat', function() {
